x64: replace PD -> PS where possible
Wherever the packed double and the packed single versions are bitwise
equivalent, the PS version is preferable because it uses one less byte
to encode (by not requiring the operand-size prefix 0x66). Recommended
by both Intel and Agner Fog.

I did this directly in the emitter because doing this in the JIT makes
it a bit confusing to read.