VTune: add headers to Externals
This way we don't have to do #ifdef acrobatics.

TODO: decide on a way to handle jitprofiling.h functions

Note that while the headers are dual-licensed under GPLv2 and New BSD,
the standard ittnotify/jitprofiling libraries themselves are not open
source, even the "open source version" at [0] is just a dlopen wrapper.
This means we cannot redistribute Dolphin binaries with VTune enabled.

[0] https://github.com/01org/IntelSEAPI