TextureDecoder: restructure/optimize
Move all the texture decoder code into its own namespace.

Remove lots of redundant code. Extract almost all the looping code into
a template. The wrappers are needed because the differing target
attributes would otherwise prevent function inlining.

Always compile all intrinsics code, not just code for the instruction
set extensions enabled on the command line. This currently requires a
hacky workaround to get it to work in Clang 3.7. As a side-effect, this
enables previously uncompiled optimizations in the audio DMA code (~1%
CPU) and in the hashing functions. MSVC supposedly already supported
this.

Optimize a lot. Most of the SSE2/SSSE3 code was not optimized at all.
Add AVX2 implementations for all formats where it makes sense. This
gives texture decoding speedups up to 500%, depending on the format.
There are a couple of FIXMEs that are unlikely to become a problem
(potential RAM overread by two bytes when using paletted texture
formats).

TODO: get it to compile on all platforms
- FreeBSD: Clang 3.4.1
- OS X: Apple LLVM version 6.0 (clang-600.0.54) (based on LLVM 3.5svn)